+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Ability to create a SmartWallet and send a user operation.
 - Add `EURC` and `cbBTC` asset support.
 
 ## [0.18.1] - 2025-02-13